Moved search results component to a separate page, updated styling
