+++ src/index.css
@@ -1,10 +1,22 @@
+* {
+    margin: 0;
+    padding: 0;
+}
+
+.body {
+    margin: 0;
+    padding: 0;
+}
+
 .app-container {
     background-image: url('./assets/sunsetFlight.png');
     background-size: cover;
-    backdrop-filter: blur(100px);
-    color: orange;
     min-height: 100vh;
     min-width: 100vw;
+    overflow-x: hidden;
+}
+
+.search-results-container {
     display: flex;
     flex-direction: column;
     align-items: center;
